footnote: builds and runs lapack package 3.10
Runs at the same speed and when used to build lapack all the lapack
tests run except the BLAS tests, which is expected.

I pulled the lapack package and changed the make(1) files to use
the libM_blas.a library instead of librefblas.a and everything built
and ran all the tests, using gfortran.

The package builds with nvfortran, ifort, and gfortran and passes
the same tests with the inclusion of a special feature to change
out xerbla which is required to run the tests with a module.

Was hoping for a speedup from more inlining, but using the tests
and running them two hundred times with timing I saw a negligle
but repeatable improvement of 0.2 percent; so no better yet but
also no worse.