checkpoint where it builds lapack without error
Interestingly, the timing on the tests in this package have not
significantly changed, but the lapack test takes 17 percent less
time when the -O2 level is used with gfortran, matching the lapack
build. Since the lapack test does a considerable amount of loading
and I have not profiled it yet, this is intriguing enough to
track down.